third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/semantics/forms/the-selectlist-element/selectlist-tabindex-order.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Check that tabindex applies to <selectlist>
  assert_equals: input3 should be active expected "input3" but got ""
Harness: the test ran to completion.
